PO-1592: Refactor Draft Accounts Controller into Service.
