+++ build.gradle
@@ -63,6 +63,8 @@ tasks.withType(JavaCompile).configureEach {
   options.compilerArgs << "-Xlint:deprecation"
 //  Enable the following to stop the build on any warning messages (not just error messages)
   options.compilerArgs << "-Werror"
+//  Change the maximum number of reported compiler errors (default 100)
+  options.compilerArgs.addAll("-Xmaxerrs", "200")
 }
 
 // https://github.com/gradle/gradle/issues/16791
@@ -314,7 +316,7 @@ dependencies {
   implementation 'com.fasterxml.jackson.core:jackson-databind:2.16.2'
   compileOnly 'org.projectlombok:lombok:1.18.36'
   annotationProcessor 'org.projectlombok:lombok:1.18.36'
-  annotationProcessor 'org.hibernate:hibernate-jpamodelgen:6.6.5.Final'
+  annotationProcessor 'org.hibernate.orm:hibernate-jpamodelgen:6.6.5.Final'
 
   testCompileOnly 'org.projectlombok:lombok:1.18.36'
   testAnnotationProcessor 'org.projectlombok:lombok:1.18.36'
